Create proper upstart job for our klogd equivalent.
The new job pipes kernel messages to rsyslog. It is a
bit nicer to have it as an upstart job rather than using
start-stop-daemon because then upstart can track the pid
and restart it as necessary. Also, I think this was the
last start-stop-daemon holdout, so we may be able to get
rid of whatever package provides that.

Review URL: http://chromereview.prom.corp.google.com/1187066


git-svn-id: svn://chrome-svn/chromeos/trunk@261 06c00378-0e64-4dae-be16-12b19f9950a1