Factory install: ability for non-factory image to wipe factory state.
The way it works: on boot, if
/mnt/stateful_partition/factory_install_reset file is detected, and
owned by root, normal startup is halted and instead clobber-state.sh
is called.

clobber-state.sh zeros out the stateful partition and re-initializes
it. Also, if booted from the second rootfs (as will be the case in the
factory), it wipes the first rootfs. It then reboots, which will cause
the machine to startup just as it would during unboxing.

We use nested ifs in chromeos-startup b/c sh doesn't do
short-circuiting for and/or statements.

TEST=tested the script on my netbook
BUG=2484

Review URL: http://codereview.chromium.org/1508032