Ensure system.data is initialized when the TPM is already owned.
This CL addresses a problem where trousers requires an initialized system.data, but that only happens fully when a successful call to Tspi_TPM_TakeOwnership is made.  If the TPM has been owned by a previous install, and isn't cleared, then trousers will not work.  To address this, the tcsd.conf file has been modified to copy a stock system.data file when it detects that the TPM is owned, but the system.data file is missing or zero-length.

See http://trousers.sourceforge.net/faq.html#1.7

Testing consists of owning the TPM, and then deleting or zero-sizing the system.data file, rebooting and running `cryptohome --action=status`, which, on success, will be able to test the TPM functionality for cryptohome.

Change-Id: I0a97fa7cd74b4dff6cdf74e8ca2d85842ff85fc3

BUG=NONE
TEST=see above

Review URL: http://codereview.chromium.org/3318011