chromeos_startup: developer mode clobber-state and delay
Part of the cros_boot_mode work, this adds two mitigating features
for enabling sudo and vt2 in dev mode.  On first detected transition:
1. sleep for 5 minutes
2. clobber-state

In addition, transitioning back to non-dev mode will clobber-state, but
it is not a reliable check as it just looks for a dotfile which the owner
could remove prior to flipping the switch.

BUG=chromium-os:824,chromium-os:6164
TEST=in progress: built x86-generic, tested on nodev and withdev on legacy fw (which means no clobber). Now testing on cros_fw to confirm sleep and clobber.

Review URL: http://codereview.chromium.org/3557002

Change-Id: I3f94b34e9ab751097391f4cab81cbfcd5c271aed