Delay light-sensor startup until I2C is available.
This fixes the problem of the light sensor driver not loading.
In order for the device to be instantiated by the script, the I2C bus
driver has to be available.

The recent move of the udev trigger from udev to udev-addon caused this problem.

BUG=chromium-os:7312
TEST=Boot, then confirm module tsl2563 is loaded.

Change-Id: I7e6b818de88ea72b4d81e637cd38ec91bcc8f515

Review URL: http://codereview.chromium.org/3964001