Make sure that the delayed transition to dev-mode is uninterruptable
We can't invoke a long-running foreground process that redirects to
/dev/tty1. If we do, it will open a controlling terminal which allows ctrl-\
to kill it before the process is done. It's okay if we kill an "echo"
occasionally, although I've been unable to prove it's possible. Further
analysis is welcomed.

Change-Id: I1bf34a577ffbd7dd5246877ca31d9272a261052f

BUG=none
TEST=manual

Switch to dev-mode. On the first boot (hit ^D or wait 30 seconds at the
scary boot screen), you should see a warning message, a countdown, and a
progress bar. You shouldn't be able to interrupt any of them.

Review URL: http://codereview.chromium.org/4039004