init: Update boot-complete to run at login-prompt-visible.
This saves us 2-3 seconds at boot.  It appears that the jobs
that were started when boot-complete finishes were greatly
slowing down Chrome while it was painting the login windows
after mapping them.  The window manager now notifies the
session manager after the windows are painted, at which
point the login-prompt-visible Upstart signal is emitted.

Here are some timings showing the seconds since power-on at
which the various signals are fired without and with this
change.

before:
login-prompt-ready     6.74   6.77   6.49
login-prompt-visible  10.69  10.61  10.21

after:
login-prompt-ready     6.37   6.78   6.65
login-prompt-visible   7.54   7.94   7.61

This also makes it write a boot-complete event (I have a
separate change to make the session manager log
login-prompt-ready and login-prompt-visible) and fixes a
typo in the logic that was supposed to prevent the job from
running a second time if the user logged out.

Change-Id: Ib11cac2ccba2a54180e1d9d8849352e1cf7b99af

BUG=chromium-os:10349
TEST=tried it.  checked that boot-complete runs and that /tmp/uptime-* files are written correctly

Review URL: http://codereview.chromium.org/5868004