Remove XKB cache files on the next boot after auto update.
The hook can be used for other tasks, that have to be
run after auto update, before X and Chrome start.

Note that impact on the boot time should be negligible as
checking "/mnt/stateful_partition/.install_completed" is
unlikely to touch the disk, as we already check existence
of files in the same directory (i.e. the file attributes
metadata is likely cached).

BUG=chromium-os:12916
TEST=confirm that /var/lib/xkb/*.xkm are removed as intended.

Change-Id: I07ced205fe44081609ebe3f30c79f350dce986c4

Review URL: http://codereview.chromium.org/6651018