init: enable support for per-user flimflam profiles
On login create the flimflam directory in cryptohome and push the user
profile. On logout pop the per-user profile. Handle the case of flimflam
crashing and being started by checking if we're logged in and explicitly
pushing on restart--this avoids doing a sleep of unknown duration after
starting flimflam to ensure it's ready for dbus cmds.

BUG=chromium-os:12948
TEST=manual:go to alt console; run list-profiles to verify only the default profile is present; login; check profiles again and verify the per-user profile is now present; logout and check again; also after kill flimflamd and verify profile stack is restored

Reviewed-on: http://gerrit.chromium.org/gerrit/1194
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
(cherry picked from commit 40594d91689a5f1ab9a315bcd3fe97dd4e2614c7)

Change-Id: Ied78cbac2e2de231b1a08305aa703e7676cac4f6
Reviewed-on: http://gerrit.chromium.org/gerrit/2215
Tested-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>