Terminate any running cryptohome processes before cryptohomed stops
The cryptohome client may be in the middle of a long pkcs#11 initialization
operation with open files on the cryptohome mount. Since this is expected,
kill it in the cryptohomed pre-stop, so that the infraction is not logged by the
killers script.

BUG=chromium-os:15697
TEST=Log in to a fresh system, immediately shutdown (or reboot from a terminal
session). Restart and make sure that /var/log/shutdown_force_kill_processes is
empty.

(cherry picked from commit b094d04baff5de88bf1f301bc0f628752dc18599)

Change-Id: Ia77eefe0b211ba99e828b08b5f87203d5b6743f3
Reviewed-on: http://gerrit.chromium.org/gerrit/2699
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>